Move visual to cmd/pics
