+++ cmd/pics/visual/common.go
