+++ cmd/pics/visual/hexary.go
