+++ cmd/pics/visual/quad.go
